Deprecated the use of -use-make used when after a run of latex or
pdflatex, there are warnings about missing files (e.g., as
requested by the LaTeX \input, \include, and \includgraphics
commands), latexmk tries to make them by a custom dependency. If
no relevant custom dependency with an appropriate source file is
found, and if the -use-make option is set, then as a last resort
latexmk will try to use the make program to try to make the
missing files.